+++ app/Http/Controllers/Backend/VocabularyController.php
@@ -166,8 +166,10 @@ class VocabularyController extends Controller
         $catId = $request->catId ? $request->catId : 0;
         $data = $request->except('_token', 'files');
         if ($request->hasFile('thumb')) {
-            $path = $request->file('thumb')->store('thumbnails', 'public');
-            $data['thumb'] = $path;
+            if (!isUrl($request->thumb)) {
+                $path = $request->file('thumb')->store('thumbnails', 'public');
+                $data['thumb'] = $path;
+            }
         }
         $isSave = $this->vocabularyRepository->storeNew($data);
         return redirect()->route('admin.vocabulary.categoryList', ['catId' => $catId])->with($isSave ? SUCCESS : ERROR, $isSave ? CREATE_SUCCESS : CREATE_ERROR);
@@ -186,16 +188,19 @@ class VocabularyController extends Controller
 
     public function categoryUpdate(VocabularyRequestUpdate $request, $catId)
     {
+        $request->request->remove('inlineRadioUpload');
         $detail = Vocabulary::find($request->id);
         $data = $request->except('_token', 'files');
         if ($request->hasFile('thumb')) {
-            $path = $request->file('thumb')->store('thumbnails', 'public');
-            $data['thumb'] = $path;
             // remove old image
-            if (!empty($detail->thumb)) {
+            if (!empty($detail->thumb) && !isUrl($request->thumb)) {
                 if (file_exists('storage/' . $detail->thumb)) {
                     unlink('storage/' . $detail->thumb);
                 };
+            }
+            if (!isUrl($request->thumb)) {
+                $path = $request->file('thumb')->store('thumbnails', 'public');
+                $data['thumb'] = $path;
             }
         }
         $isSave = $this->vocabularyRepository->update($request->id, $data);
